Create Restaurant.py
Restaurant
Prepbudy has opened a restaurant recently. He received an order to prepare N dishes. Dish i can be cooked by using an oven for Ai consecutive minutes. 
If Prepbudy has only a2 oven in the restaurant to use. what is the shortest number of minutes needed to cook all the N dishes?

Note: Assume that all the dishes need an oven and only one dish can be cooked in the oven at one time.

Input format
The first line of the input contains a single integer T - the number of test cases.
The first line of each test case contains a single integer N - the number of dishes.
The second line of each test contains N space-separated integers A1,A2....An - the number of minutes required to cook the ith dish.

Output format
For each test case, print a single integer - the shortest number of minutes needed to cook all the N dishes.

Constraints
1≤T≤100
1≤N≤2∗10^2
1≤Ai≤2∗10^3

Time Limit
1 second

Example
Sample Input
1
5
8 3 7 2 5

Sample Output
13

Sample Test Case Explanation
For this test case, if we use the ovens in the following manners, we can cook all the 5 dishes in 13 minutes.

The first oven: Cook Dishes  5 and 1 in this order.
The second oven: Cook Dishes  2,  4, and 3 in this order.